docs(guides): minor fixes and typos (#4419)
* fix: consistent punctuation in list

two of these have periods and two don't... so I looked around at what is done elsewhere and took my best guess

* adds linkback to docs for refetchOnWindowFocus

at this point in the readthrough, this is the very first time I'm seeing anything about "options" and having a link to one is nice, just to quickly see what they look like and hop right back to this page.  I'd love to add some docs to the 4 that are mentioned in the next paragraph but they don't seem to have a way to link to them since they're contained in an unordered list on the useQuery page

* moves asides to be in context with attached bullets

I don't see how to run the docs site locally to test this, but at least on GitHub's markdown parser this would really help align the quote block with the associated bullet (as it is today is sorta hard to follow because some are global and some are attached to bullets, yet they're all at the same level).

feel free to drop this commit if it doesn't work for the actual markdown parser.

* adds missing inline codeblock for signal obj

* matches QueryFunctionContext types to source code

see https://github.com/TanStack/query/blob/ee9295c7dd3d2ea7f135ff46fa0c27038b2eda83/packages/query-core/src/types.ts#L24

following the standard of https://www.typescriptlang.org/tsconfig#exactOptionalPropertyTypes these types are subtly incorrect.

* fix multiple h1s on page (and this one pushing above)

check out https://tanstack.com/query/v4/docs/guides/background-fetching-indicators#displaying-global-background-fetching-loading-state

and see that the section that's an h1 at the bottom is pushed pretty hard up against the codeblock before it.  I think this is some kind of tailwind problem, but it's a win-win because a page having multiple h1 elements is an accessibility nit anyways